requirements.txt
flask==3.0.3
gunicorn==22.0.0

requests==2.32.3
yfinance==0.2.43
Pillow==10.4.0

google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
importlib-metadata==8.2.0